+++ CHANGELOG.rst
@@ -24,6 +24,9 @@ Release 2.1.9 (in development)
 * Added support for BSPackedAdditionalGeometryData (reported by
   Ghostwalker71, niftools issue #3177847).
 
+* Skip terminal chars in mopper output (fixes issues with running
+  mopper under wine).
+
 Release 2.1.8 (4 February 2011)
 ===============================